+++ sources_rec/views.py
@@ -4,6 +4,7 @@ from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
 from sources_rec import service
 from sources_rec import models
+from sources_rec import validators
 
 
 @api_view(["POST"])
@@ -45,7 +46,7 @@ def add_source(request):
         "ratings_5": 0
     }
     try:
-        service.validate_source_input(source)
+        validators.validate_source_input(source)
         service.add_source(source)
     except ValidationError:
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -63,7 +64,7 @@ def get_sources(request):
 def get_source(request):
     url = request.data.get("url")
     try:
-        service.validate_get_source_input(url)
+        validators.validate_get_source_input(url)
         source = service.get_source(url)
     except (ValidationError, models.Source.DoesNotExist):
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -77,7 +78,7 @@ def delete_source(request):
         "id": request.data.get("id")
     }
     try:
-        service.validate_delete_source_input(source)
+        validators.validate_delete_source_input(source)
         service.delete_source(source)
     except (ValidationError, models.Source.DoesNotExist):
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -94,7 +95,7 @@ def update_source(request):
         "url": request.data.get("url"),
     }
     try:
-        service.validate_source_input(source)
+        validators.validate_source_input(source)
         service.update_source(source)
     except ValidationError:
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -110,7 +111,7 @@ def add_rating(request):
         "rating": request.data.get("rating"),
     }
     try:
-        service.validate_rating_input(rating)
+        validators.validate_rating_input(rating)
         service.add_rating(rating)
     except (ValidationError, models.Source.DoesNotExist):
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -122,7 +123,7 @@ def add_rating(request):
 def get_user_ratings(request):
     user = request.data.get("user")
     try:
-        service.validate_user_or_source_ratings_input(user)
+        validators.validate_user_or_source_ratings_input(user)
         ratings = service.get_user_ratings(user)
     except ValidationError:
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -134,7 +135,7 @@ def get_user_ratings(request):
 def get_source_ratings(request):
     source = request.data.get("source")
     try:
-        service.validate_user_or_source_ratings_input(source)
+        validators.validate_user_or_source_ratings_input(source)
         ratings = service.get_source_ratings(source)
     except ValidationError:
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -147,8 +148,8 @@ def get_source_and_user_ratings(request):
     source = request.data.get("source")
     user = request.data.get("user")
     try:
-        service.validate_user_or_source_ratings_input(source)
-        service.validate_user_or_source_ratings_input(user)
+        validators.validate_user_or_source_ratings_input(source)
+        validators.validate_user_or_source_ratings_input(user)
         ratings = service.get_source_and_user_ratings(source, user)
     except ValidationError:
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -163,7 +164,7 @@ def delete_rating(request):
         "user": request.data.get("user"),
     }
     try:
-        service.validate_delete_rating_input(rating)
+        validators.validate_delete_rating_input(rating)
         service.delete_rating(rating)
     except (ValidationError, models.Rating.DoesNotExist):
         return Response(status=status.HTTP_400_BAD_REQUEST)
@@ -179,7 +180,7 @@ def update_rating(request):
         "rating": request.data.get("rating"),
     }
     try:
-        service.validate_rating_input(rating)
+        validators.validate_rating_input(rating)
         service.update_rating(rating)
     except (ValidationError, models.Rating.DoesNotExist):
         return Response(status=status.HTTP_400_BAD_REQUEST)
